+++ src/server/actions/meeting/archive/action.ts
@@ -1,44 +1,11 @@
 "use server";
 
-import { eq } from "drizzle-orm";
-import { db } from "@/db";
-import { meetings, type SelectMeeting } from "@/db/schema";
+import type { SelectMeeting } from "@/db/schema";
 import { getCurrentSession } from "@/lib/auth";
-import type { MeetingMemberActionResult } from "@/server/actions/meeting/leave/action";
-
-/**
- * Archives ("deletes") a meeting on behalf of `memberId`, who must be its
- * host. The session-reading `archiveMeeting` below and
- * `POST /api/meetings/[id]/archive` both end up here.
- */
-export async function archiveMeetingForMember(
-	meetingId: string,
-	memberId: string,
-): Promise<MeetingMemberActionResult> {
-	const [meeting] = await db
-		.select({ hostId: meetings.hostId })
-		.from(meetings)
-		.where(eq(meetings.id, meetingId))
-		.limit(1);
-
-	if (!meeting) {
-		return { success: false, error: "Meeting not found." };
-	}
-
-	if (meeting.hostId !== memberId) {
-		return {
-			success: false,
-			error: "Only meeting owner has permission to delete this meeting.",
-		};
-	}
-
-	await db
-		.update(meetings)
-		.set({ archived: true })
-		.where(eq(meetings.id, meetingId));
-
-	return { success: true };
-}
+import {
+	archiveMeetingForMember,
+	type MeetingMemberActionResult,
+} from "@/server/data/meeting/member-actions";
 
 export async function archiveMeeting(
 	meetingData: SelectMeeting,
